OBP上mlaprolog v3支持GLM(hcq支持2048，d支持192)



# message auto-generated for no-merge-commit merge:
!2801 merge master into master

OBP上mlaprolog v3支持GLM(hcq支持2048，d支持192)

Created-by: taomengyuan-hw
Commit-by: taomengyuan-hw
Merged-by: cann-robot
Description: ## 描述
OBP上mlaprolog v3支持GLM(hcq支持2048，d支持192)

## 关联的Issue
https://gitcode.com/cann/ops-transformer/issues/1306
## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于二级冒烟、算子泛化等。-->

## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 类型标签
<!--  [x] 表示选中 -->
- [ ] 🐛 Bug 修复
- [x] ✨ 新特性
- [ ] ⚡ 性能优化
- [ ] ♻️ 重构
- [ ] 🧪 测试
- [ ] 📦 构建/CI
- [ ] 🔧 配置变更
- [ ] 📝 文档更新
- [ ] ⬆️ 依赖升级
- [ ] 🔒 安全修复
- [ ] 🧹 代码清理
- [ ] ❓ 其他，请描述：



See merge request: cann/ops-transformer!2801

+++ attention/mla_prolog/op_host/mla_prolog_tiling.h
@@ -287,12 +287,12 @@ struct MlaPrologContext {
     const char *cacheMode;
     const bool *queryNormFlag;
 
-    const int *weightQuantMode;
-    const int *kvQuantMode;
-    const int *queryQuantMode;
-    const int *ckvkrRepoMode;
-    const int *quantScaleRepoMode;
-    const int *tileSize;
+    const int64_t *weightQuantMode;
+    const int64_t *kvQuantMode;
+    const int64_t *queryQuantMode;
+    const int64_t *ckvkrRepoMode;
+    const int64_t *quantScaleRepoMode;
+    const int64_t *tileSize;
 
     const float *qcQrScale;
     const float *kcScale;
